fix(m10-03): close yardage book loop and resolve review gaps
Persist optional target_yardage on analysis create, add params UI and profile entry, document yardage-book APIs in docs/02, fix ruff in yardage_inference, and route yardage edits to the clubs page.

+++ backend/app/services/yardage_inference.py
@@ -43,6 +43,6 @@ async def infer_yardage_for_club(
     rows = [int(r) for r in (await db.execute(stmt)).scalars().all() if r is not None]
     if len(rows) < MIN_INFERENCE_SAMPLES:
         return None
-    avg = int(round(statistics.mean(rows)))
+    avg = round(statistics.mean(rows))
     std = float(statistics.pstdev(rows)) if len(rows) > 1 else 0.0
     return YardageInference(avg=avg, std=std, sample_count=len(rows))
